Remove all for loops
Here we are using again forEach but forEach call callback function inside it. his callback function is not directly method of view object. So in this case when we execute teh code we get this.createDeleteButton is not a function error we need to specify tis statement as a parameter for forEach statement.This is used as a parameter in the forEach